+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Get-QualysTags
+- DefaultScanner parameter for add/set Qualys Asset Groups
 
 ### Changed